+++ CHANGES.rst
@@ -7,6 +7,7 @@
   `~asdf.search.AsdfSearchResult.schema_info` method. [#1197]
 - Use forc ndarray flag to correctly test for fortran array contiguity [#1206]
 - Unpin ``jsonschema`` version and fix ``jsonschema`` deprecation warnings. [#1185]
+- Replace ``pkg_resources`` with ``importlib.metadata``. [#1199]
 
 2.13.0 (2022-08-19)
 -------------------